requirements.txt
numpy
scipy
matplotlib
scikit-learn
ipykernel
pandas
transformers
torch
accelerate
python-dotenv
